(feat) integrated OCI AI
